# openapi.yaml
openapi: 3.0.3
info:
  title: Carbon Travel Intelligence API
  description: |
    **"Stripe for sustainability data in travel"**
    
    A Carbon- & Resource-Aware Travel Intelligence Platform providing real-time 
    carbon impact calculations, lower-impact alternatives, and CSRD-aligned ESG outputs.
    
    ## Core Capabilities
    - **Real Carbon Impact per Trip**: Flight emissions, ground transport, hotel energy intensity
    - **"Better Alternative" Engine**: AI-powered suggestions for lower-emission itineraries  
    - **Audit-Ready ESG Outputs**: CSRD-aligned exports with transparent calculations
    
    ## MVP Scope
    - Flights + Hotels only
    - EU-first regional grid carbon intensity data
    - REST API with JSON responses
    
  version: 1.0.0
  contact:
    name: Carbon Travel API Support
    email: api@carbontravelintel.io
  license:
    name: Proprietary
    url: https://carbontravelintel.io/terms
servers:
  - url: https://api.carbontravelintel.io/v1
    description: Production
  - url: https://sandbox.carbontravelintel.io/v1
    description: Sandbox (test data)
...
paths:
...
  /factors/flights:
    get:
      operationId: getFlightEmissionFactors
      summary: Get flight emission factors
      description: |
        Returns current emission factors used for flight calculations.
        Based on ICAO Carbon Emissions Calculator methodology.
      tags:
        - Carbon Assessment
      security:
        - ApiKeyAuth: []
      parameters:
        - name: cabin_class
          in: query
          schema:
            type: string
            enum: [economy, premium_economy, business, first]
      responses:
        '200':
          description: Flight emission factors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightFactorsResponse'
  /factors/hotels:
    get:
      operationId: getHotelEmissionFactors
      summary: Get hotel emission factors by region
      description: |
        Returns regional hotel emission factors including grid carbon intensity.
        EU regions have granular data; other regions use IPCC defaults.
      tags:
        - Carbon Assessment
      security:
        - ApiKeyAuth: []
      parameters:
        - name: country_code
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 2
          example: "FR"
      responses:
        '200':
          description: Hotel emission factors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelFactorsResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    FlightFactorsResponse:
      type: object
      properties:
        factors:
          type: array
          items:
            type: object
            properties:
              cabin_class:
                type: string
              haul_type:
                type: string
              kg_co2e_per_km:
                type: number
                format: double
              radiative_forcing_multiplier:
                type: number
                format: double
        source:
          type: string
        updated_at:
          type: string
          format: date
    HotelFactorsResponse:
      type: object
      properties:
        country_code:
          type: string
        country_name:
          type: string
        grid_carbon_intensity:
          type: object
          properties:
            value_g_co2_per_kwh:
              type: number
              format: double
            source:
              type: string
            year:
              type: integer
        hotel_emission_factors:
          type: array
          items:
            type: object
            properties:
              star_rating:
                type: integer
              kg_co2e_per_night:
                type: number
                format: double
              kwh_per_night:
                type: number
                format: double
        data_quality:
          type: string
          enum: [measured, estimated, default]
          description: |
            - measured: Country-specific real data
            - estimated: Regional estimates
            - default: IPCC global defaults
...